- initial CFData implementation - NSBitmapImageRep -TIFFRepresentationOfImageRepsInArray:, -TIFFRepresentationOfImageRepsInArray:usingCompression:factor - NSImage -TIFFRepresentation, -TIFFRepresentationUsingCompression:factor: - renamed NSTIFFReader to O2Decoder_TIFF - renamed O2ColorSpaceType constants to kO2ColorSpaceModel ones - initial CG/O2ImageDestination implementation supporting simple TIFF encoding via O2Encoder_TIFF - O2DataProvider, O2DataConsumer work - Some CF related cleanup in O2/CG

+++ CoreGraphics/O2Font_ttf.m
@@ -13,7 +13,7 @@
 }
 
 -(O2Glyph)glyphWithGlyphName:(NSString *)name {
-   return (O2Glyph)NSMapGet(_nameToGlyph,name);
+   return (O2Glyph)(int)NSMapGet(_nameToGlyph,name);
 }
 
 @end
